+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaConnectBuildTest.java
@@ -48,6 +48,7 @@ public class KafkaConnectBuildTest {
 
     private final String cluster = "my-connect";
     private final String namespace = "my-ns";
+    private final boolean useBuildah = false;
 
     private final Artifact jarArtifactNoChecksum = new JarArtifactBuilder()
             .withUrl("https://mydomain.tld/my.jar")
@@ -58,9 +59,12 @@ public class KafkaConnectBuildTest {
             .withSha512sum("sha-512-checksum")
             .build();
 
-    private final List<String> defaultArgs = List.of("--dockerfile=/dockerfile/Dockerfile",
+    private final List<String> defaultKanikoArgs = List.of("--dockerfile=/dockerfile/Dockerfile",
             "--image-name-with-digest-file=/dev/termination-log",
             "--destination=my-image:latest");
+
+    private final String defaultBuildahBuildArgs = "--file=/dockerfile/Dockerfile --tag=my-image:latest --storage-driver=vfs";
+    private final String defaultBuildahPushArgs = "--storage-driver=vfs --digestfile=/tmp/digest";
 
     @Test
     public void testFromCrd()   {
@@ -82,7 +86,7 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah);
     }
 
     @Test
@@ -104,7 +108,7 @@ public class KafkaConnectBuildTest {
                 .build();
 
         assertThrows(InvalidResourceException.class, () ->
-            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER)
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah)
         );
     }
 
@@ -128,7 +132,7 @@ public class KafkaConnectBuildTest {
                 .build();
 
         assertThrows(InvalidResourceException.class, () ->
-            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER)
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah)
         );
     }
 
@@ -153,12 +157,12 @@ public class KafkaConnectBuildTest {
                 .build();
 
         assertThrows(InvalidResourceException.class, () ->
-            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER)
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah)
         );
     }
 
     @Test
-    public void testDeployment()   {
+    public void testKanikoDeployment()   {
         Map<String, Quantity> limit = new HashMap<>();
         limit.put("cpu", new Quantity("500m"));
         limit.put("memory", new Quantity("512Mi"));
@@ -187,11 +191,11 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, false);
 
         assertThat(build.baseImage, is("my-source-image:latest"));
 
-        Pod pod = build.generateBuilderPod(true, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        Pod pod = build.generateBuilderPod(true, false, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
         assertThat(pod.getMetadata().getName(), is(KafkaConnectResources.buildPodName(cluster)));
         assertThat(pod.getMetadata().getNamespace(), is(namespace));
 
@@ -206,9 +210,10 @@ public class KafkaConnectBuildTest {
         assertThat(pod.getMetadata().getLabels(), is(expectedDeploymentLabels));
         assertThat(pod.getSpec().getServiceAccountName(), is(KafkaConnectResources.buildServiceAccountName(cluster)));
         assertThat(pod.getSpec().getContainers().size(), is(1));
-        assertThat(pod.getSpec().getContainers().get(0).getArgs(), is(defaultArgs));
+        assertThat(pod.getSpec().getContainers().get(0).getArgs(), is(defaultKanikoArgs));
         assertThat(pod.getSpec().getContainers().get(0).getName(), is(KafkaConnectResources.buildPodName(this.cluster)));
-        assertThat(pod.getSpec().getContainers().get(0).getImage(), is(build.image));
+        // TODO: use configuration from the `ClusterOperatorConfig` rather than from env variables directly - https://github.com/strimzi/strimzi-kafka-operator/issues/11981
+        assertThat(pod.getSpec().getContainers().get(0).getImage(), is(KafkaConnectBuild.DEFAULT_KANIKO_EXECUTOR_IMAGE));
         assertThat(pod.getSpec().getContainers().get(0).getPorts(), is(nullValue()));
         assertThat(pod.getSpec().getContainers().get(0).getResources().getLimits(), is(limit));
         assertThat(pod.getSpec().getContainers().get(0).getResources().getRequests(), is(request));
@@ -226,7 +231,79 @@ public class KafkaConnectBuildTest {
     }
 
     @Test
-    public void testDeploymentWithoutPushSecret()   {
+    public void testBuildahDeployment() {
+        Map<String, Quantity> limit = new HashMap<>();
+        limit.put("cpu", new Quantity("500m"));
+        limit.put("memory", new Quantity("512Mi"));
+
+        Map<String, Quantity> request = new HashMap<>();
+        request.put("cpu", new Quantity("1000m"));
+        request.put("memory", new Quantity("1Gi"));
+
+        KafkaConnect kc = new KafkaConnectBuilder()
+            .withNewMetadata()
+                .withName(cluster)
+                .withNamespace(namespace)
+            .endMetadata()
+            .withNewSpec()
+                .withImage("my-source-image:latest")
+                .withBootstrapServers("my-kafka:9092")
+                .withNewBuild()
+                    .withNewDockerOutput()
+                        .withImage("my-image:latest")
+                        .withPushSecret("my-docker-credentials")
+                    .endDockerOutput()
+                    .withPlugins(new PluginBuilder().withName("my-connector").withArtifacts(jarArtifactWithChecksum).build(),
+                        new PluginBuilder().withName("my-connector2").withArtifacts(jarArtifactNoChecksum).build())
+                    .withResources(new ResourceRequirementsBuilder().withLimits(limit).withRequests(request).build())
+                .endBuild()
+            .endSpec()
+            .build();
+
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, true);
+
+        assertThat(build.baseImage, is("my-source-image:latest"));
+
+        Pod pod = build.generateBuilderPod(true, true, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        assertThat(pod.getMetadata().getName(), is(KafkaConnectResources.buildPodName(cluster)));
+        assertThat(pod.getMetadata().getNamespace(), is(namespace));
+
+        Map<String, String> expectedDeploymentLabels = Map.of(Labels.STRIMZI_CLUSTER_LABEL, this.cluster,
+            Labels.STRIMZI_NAME_LABEL, KafkaConnectResources.buildPodName(cluster),
+            Labels.STRIMZI_KIND_LABEL, KafkaConnect.RESOURCE_KIND,
+            Labels.STRIMZI_COMPONENT_TYPE_LABEL, KafkaConnectBuild.COMPONENT_TYPE,
+            Labels.KUBERNETES_NAME_LABEL, KafkaConnectBuild.COMPONENT_TYPE,
+            Labels.KUBERNETES_INSTANCE_LABEL, this.cluster,
+            Labels.KUBERNETES_PART_OF_LABEL, Labels.APPLICATION_NAME + "-" + this.cluster,
+            Labels.KUBERNETES_MANAGED_BY_LABEL, AbstractModel.STRIMZI_CLUSTER_OPERATOR_NAME);
+        String[] commands = pod.getSpec().getContainers().get(0).getArgs().get(2).split("\n");
+
+        assertThat(pod.getMetadata().getLabels(), is(expectedDeploymentLabels));
+        assertThat(pod.getSpec().getServiceAccountName(), is(KafkaConnectResources.buildServiceAccountName(cluster)));
+        assertThat(pod.getSpec().getContainers().size(), is(1));
+        assertThat(commands[0], containsString(defaultBuildahBuildArgs));
+        assertThat(commands[1], containsString(defaultBuildahPushArgs));
+        assertThat(pod.getSpec().getContainers().get(0).getName(), is(KafkaConnectResources.buildPodName(this.cluster)));
+        // TODO: use configuration from the `ClusterOperatorConfig` rather than from env variables directly - https://github.com/strimzi/strimzi-kafka-operator/issues/11981
+        assertThat(pod.getSpec().getContainers().get(0).getImage(), is(KafkaConnectBuild.DEFAULT_BUILDAH_IMAGE));
+        assertThat(pod.getSpec().getContainers().get(0).getPorts(), is(nullValue()));
+        assertThat(pod.getSpec().getContainers().get(0).getResources().getLimits(), is(limit));
+        assertThat(pod.getSpec().getContainers().get(0).getResources().getRequests(), is(request));
+        assertThat(pod.getSpec().getVolumes().size(), is(2));
+        assertThat(pod.getSpec().getVolumes().get(0).getName(), is("dockerfile"));
+        assertThat(pod.getSpec().getVolumes().get(0).getConfigMap().getName(), is(KafkaConnectResources.dockerFileConfigMapName(cluster)));
+        assertThat(pod.getSpec().getVolumes().get(1).getName(), is("docker-credentials"));
+        assertThat(pod.getSpec().getVolumes().get(1).getSecret().getSecretName(), is("my-docker-credentials"));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().size(), is(2));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().get(0).getName(), is("dockerfile"));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().get(0).getMountPath(), is("/dockerfile"));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().get(1).getName(), is("docker-credentials"));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().get(1).getMountPath(), is("/kaniko/.docker"));
+        io.strimzi.operator.cluster.TestUtils.checkOwnerReference(pod, kc);
+    }
+
+    @Test
+    public void testKanikoDeploymentWithoutPushSecret()   {
         KafkaConnect kc = new KafkaConnectBuilder()
                 .withNewMetadata()
                     .withName(cluster)
@@ -244,11 +321,45 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, false);
 
-        Pod pod = build.generateBuilderPod(true, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        Pod pod = build.generateBuilderPod(true, false, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
         assertThat(pod.getSpec().getVolumes().size(), is(1));
-        assertThat(pod.getSpec().getContainers().get(0).getArgs(), is(defaultArgs));
+        assertThat(pod.getSpec().getContainers().get(0).getArgs(), is(defaultKanikoArgs));
+        assertThat(pod.getSpec().getVolumes().get(0).getName(), is("dockerfile"));
+        assertThat(pod.getSpec().getVolumes().get(0).getConfigMap().getName(), is(KafkaConnectResources.dockerFileConfigMapName(cluster)));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().size(), is(1));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().get(0).getName(), is("dockerfile"));
+        assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().get(0).getMountPath(), is("/dockerfile"));
+    }
+
+    @Test
+    public void testBuildahDeploymentWithoutPushSecret()   {
+        KafkaConnect kc = new KafkaConnectBuilder()
+            .withNewMetadata()
+                .withName(cluster)
+                .withNamespace(namespace)
+            .endMetadata()
+            .withNewSpec()
+                .withBootstrapServers("my-kafka:9092")
+                .withNewBuild()
+                    .withNewDockerOutput()
+                        .withImage("my-image:latest")
+                    .endDockerOutput()
+                    .withPlugins(new PluginBuilder().withName("my-connector").withArtifacts(jarArtifactWithChecksum).build(),
+                        new PluginBuilder().withName("my-connector2").withArtifacts(jarArtifactNoChecksum).build())
+                .endBuild()
+            .endSpec()
+            .build();
+
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, true);
+
+        Pod pod = build.generateBuilderPod(true, true, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        String[] commands = pod.getSpec().getContainers().get(0).getArgs().get(2).split("\n");
+
+        assertThat(pod.getSpec().getVolumes().size(), is(1));
+        assertThat(commands[0], containsString(defaultBuildahBuildArgs));
+        assertThat(commands[1], containsString(defaultBuildahPushArgs));
         assertThat(pod.getSpec().getVolumes().get(0).getName(), is("dockerfile"));
         assertThat(pod.getSpec().getVolumes().get(0).getConfigMap().getName(), is(KafkaConnectResources.dockerFileConfigMapName(cluster)));
         assertThat(pod.getSpec().getContainers().get(0).getVolumeMounts().size(), is(1));
@@ -276,7 +387,7 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah);
 
         KafkaConnectDockerfile dockerfile = new KafkaConnectDockerfile("my-image:latest", kc.getSpec().getBuild(), SHARED_ENV_PROVIDER);
         ConfigMap cm = build.generateDockerfileConfigMap(dockerfile);
@@ -316,7 +427,7 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah);
 
         KafkaConnectDockerfile dockerfile = new KafkaConnectDockerfile("my-image:latest", kc.getSpec().getBuild(), SHARED_ENV_PROVIDER);
         BuildConfig bc = build.generateBuildConfig(dockerfile);
@@ -374,7 +485,7 @@ public class KafkaConnectBuildTest {
                 .build();
 
         InvalidResourceException thrown = assertThrows(InvalidResourceException.class, () -> {
-            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah);
         }, "InvalidResourceException was expected");
         assertThat(thrown.getMessage(), is("KafkaConnect .spec.image cannot be the same as .spec.build.output.image"));
     }
@@ -398,7 +509,7 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, false);
 
         KafkaConnectDockerfile dockerfile = new KafkaConnectDockerfile("my-image:latest", kc.getSpec().getBuild(), SHARED_ENV_PROVIDER);
         BuildConfig bc = build.generateBuildConfig(dockerfile);
@@ -494,9 +605,9 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, useBuildah);
 
-        Pod pod = build.generateBuilderPod(true, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        Pod pod = build.generateBuilderPod(true, useBuildah, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
         assertThat(pod.getMetadata().getLabels().entrySet().containsAll(buildPodLabels.entrySet()), is(true));
         assertThat(pod.getMetadata().getAnnotations().entrySet().containsAll(buildPodAnnos.entrySet()), is(true));
         assertThat(pod.getSpec().getPriorityClassName(), is("top-priority"));
@@ -520,7 +631,7 @@ public class KafkaConnectBuildTest {
 
     @Test
     public void testValidKanikoOptions()   {
-        List<String> expectedArgs = new ArrayList<>(defaultArgs);
+        List<String> expectedArgs = new ArrayList<>(defaultKanikoArgs);
         expectedArgs.add("--reproducible");
         expectedArgs.add("--single-snapshot");
         expectedArgs.add("--log-format=json");
@@ -544,9 +655,9 @@ public class KafkaConnectBuildTest {
                 .endSpec()
                 .build();
 
-        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER);
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, false);
 
-        Pod pod = build.generateBuilderPod(true, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        Pod pod = build.generateBuilderPod(true, false, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
         assertThat(pod.getSpec().getContainers().get(0).getArgs(), is(expectedArgs));
     }
 
@@ -572,9 +683,161 @@ public class KafkaConnectBuildTest {
                 .build();
 
         InvalidResourceException e = assertThrows(InvalidResourceException.class, () ->
-            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER)
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, false)
         );
 
-        assertThat(e.getMessage(), containsString(".spec.build.additionalKanikoOptions contains forbidden options: [--reproducible-something, --build-arg, --digest-file]"));
+        assertThat(e.getMessage(), containsString(".spec.build.output.additionalKanikoOptions contains forbidden options: [--reproducible-something, --build-arg, --digest-file]"));
+    }
+
+    @Test
+    public void testValidBuildahBuildOptions() {
+        String expectedBuildOptions = defaultBuildahBuildArgs + " --retry=3";
+        String expectedPushOptions = defaultBuildahPushArgs + " --quiet";
+
+        KafkaConnect kc = new KafkaConnectBuilder()
+            .withNewMetadata()
+                .withName(cluster)
+                .withNamespace(namespace)
+            .endMetadata()
+            .withNewSpec()
+                .withBootstrapServers("my-kafka:9092")
+                .withNewBuild()
+                    .withNewDockerOutput()
+                        .withImage("my-image:latest")
+                        .withPushSecret("my-docker-credentials")
+                        .withAdditionalBuildOptions("--retry=3")
+                        .withAdditionalPushOptions("--quiet")
+                    .endDockerOutput()
+                    .withPlugins(new PluginBuilder().withName("my-connector").withArtifacts(jarArtifactWithChecksum).build(),
+                        new PluginBuilder().withName("my-connector2").withArtifacts(jarArtifactNoChecksum).build())
+                .endBuild()
+            .endSpec()
+            .build();
+
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, true);
+
+        Pod pod = build.generateBuilderPod(true, true, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        String[] commands = pod.getSpec().getContainers().get(0).getArgs().get(2).split("\n");
+        assertThat(commands[0], containsString(expectedBuildOptions));
+        assertThat(commands[1], containsString(expectedPushOptions));
+    }
+
+    @Test
+    public void testInvalidBuildahBuildOptions() {
+        KafkaConnect kc = new KafkaConnectBuilder()
+            .withNewMetadata()
+                .withName(cluster)
+                .withNamespace(namespace)
+            .endMetadata()
+            .withNewSpec()
+                .withBootstrapServers("my-kafka:9092")
+                .withNewBuild()
+                    .withNewDockerOutput()
+                        .withImage("my-image:latest")
+                        .withPushSecret("my-docker-credentials")
+                        .withAdditionalBuildOptions("--logfile=my-file", "--file=/docker/Another", "--storage-driver=random")
+                        .withAdditionalPushOptions("--quiet")
+                    .endDockerOutput()
+                    .withPlugins(new PluginBuilder().withName("my-connector").withArtifacts(jarArtifactWithChecksum).build(),
+                        new PluginBuilder().withName("my-connector2").withArtifacts(jarArtifactNoChecksum).build())
+                .endBuild()
+            .endSpec()
+            .build();
+
+        InvalidResourceException e = assertThrows(InvalidResourceException.class, () ->
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, true)
+        );
+
+        assertThat(e.getMessage(), containsString(".spec.build.output.additionalBuildOptions contains forbidden options: [--logfile, --file, --storage-driver]"));
+    }
+
+    @Test
+    public void testInvalidBuildahPushOptions() {
+        KafkaConnect kc = new KafkaConnectBuilder()
+            .withNewMetadata()
+                .withName(cluster)
+                .withNamespace(namespace)
+            .endMetadata()
+            .withNewSpec()
+                .withBootstrapServers("my-kafka:9092")
+                .withNewBuild()
+                    .withNewDockerOutput()
+                        .withImage("my-image:latest")
+                        .withPushSecret("my-docker-credentials")
+                        .withAdditionalBuildOptions("--retry=3")
+                        .withAdditionalPushOptions("--digestfile=/tmp/random", "--sign-by=different", "--format=xy")
+                    .endDockerOutput()
+                    .withPlugins(new PluginBuilder().withName("my-connector").withArtifacts(jarArtifactWithChecksum).build(),
+                        new PluginBuilder().withName("my-connector2").withArtifacts(jarArtifactNoChecksum).build())
+                .endBuild()
+            .endSpec()
+            .build();
+
+        InvalidResourceException e = assertThrows(InvalidResourceException.class, () ->
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, true)
+        );
+
+        assertThat(e.getMessage(), containsString(".spec.build.output.additionalPushOptions contains forbidden options: [--digestfile, --sign-by, --format]"));
+    }
+
+    @Test
+    public void testKanikoAdditionalOptionsConfiguredOnBothPlaces() {
+        List<String> expectedArgs = new ArrayList<>(defaultKanikoArgs);
+        expectedArgs.add("--single-snapshot");
+        expectedArgs.add("--log-format=json");
+
+        KafkaConnect kc = new KafkaConnectBuilder()
+            .withNewMetadata()
+                .withName(cluster)
+                .withNamespace(namespace)
+            .endMetadata()
+            .withNewSpec()
+                .withBootstrapServers("my-kafka:9092")
+                .withNewBuild()
+                    .withNewDockerOutput()
+                        .withImage("my-image:latest")
+                        .withPushSecret("my-docker-credentials")
+                        .withAdditionalKanikoOptions("--reproducible")
+                        .withAdditionalBuildOptions("--single-snapshot", "--log-format=json")
+                    .endDockerOutput()
+                    .withPlugins(new PluginBuilder().withName("my-connector").withArtifacts(jarArtifactWithChecksum).build(),
+                        new PluginBuilder().withName("my-connector2").withArtifacts(jarArtifactNoChecksum).build())
+                .endBuild()
+            .endSpec()
+            .build();
+
+        KafkaConnectBuild build = KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, false);
+
+        Pod pod = build.generateBuilderPod(true, false, ImagePullPolicy.IFNOTPRESENT, null, "cf065b80ede090aa");
+        assertThat(pod.getSpec().getContainers().get(0).getArgs(), is(expectedArgs));
+        assertThat(pod.getSpec().getContainers().get(0).getArgs().contains("--reproducible"), is(false));
+    }
+
+    @Test
+    public void testInvalidKanikoOptionsInAdditionalBuildOptions() {
+        KafkaConnect kc = new KafkaConnectBuilder()
+            .withNewMetadata()
+                .withName(cluster)
+                .withNamespace(namespace)
+            .endMetadata()
+            .withNewSpec()
+                .withBootstrapServers("my-kafka:9092")
+                .withNewBuild()
+                    .withNewDockerOutput()
+                        .withImage("my-image:latest")
+                        .withPushSecret("my-docker-credentials")
+                        .withAdditionalBuildOptions("--reproducible", "--reproducible-something", "--build-arg", "--single-snapshot", "--digest-file=/dev/null", "--log-format=json")
+                    .endDockerOutput()
+                    .withPlugins(new PluginBuilder().withName("my-connector").withArtifacts(jarArtifactWithChecksum).build(),
+                        new PluginBuilder().withName("my-connector2").withArtifacts(jarArtifactNoChecksum).build())
+                .endBuild()
+            .endSpec()
+            .build();
+
+        InvalidResourceException e = assertThrows(InvalidResourceException.class, () ->
+            KafkaConnectBuild.fromCrd(new Reconciliation("test", kc.getKind(), kc.getMetadata().getNamespace(), kc.getMetadata().getName()), kc, VERSIONS, SHARED_ENV_PROVIDER, false)
+        );
+
+        assertThat(e.getMessage(), containsString(".spec.build.output.additionalBuildOptions contains forbidden options: [--reproducible-something, --build-arg, --digest-file]"));
     }
 }